much more robust voxelization algorithm
Before, I was using a modified even-odd rule that did not work well in
the face of stray triangles (e.g. an extra triangle just sitting inside
the mesh, breaking the even-odd rule).

The new algorithm is technically correct in the limit. A point is
"inside" the mesh if it is not reachable via any path from outside the
mesh that does not pass through a triangle in the mesh. To figure out
which points are inside the mesh, I perform breadth-first search over
the entire voxel space, calling two voxels "connected" if a line going
between their centers does not hit the mesh. This may miss small holes,
but that might actually be a good thing for crappy, hole-filled meshes.

Some meshes might have no volume and actually might have no orientation
(e.g. a klein bottle), in which case we'd like to construct voxels near
the surface to emulate the surface with a volume. Thus, my algorithm
also fills in voxels when two voxels cannot touch due to a mesh
collision, filling in the faces of the mesh with thicker voxels.